+++ src/main/kotlin/co/yappuworld/global/config/FcmConfig.kt
@@ -7,8 +7,10 @@ import com.google.firebase.FirebaseOptions
 import com.google.firebase.messaging.FirebaseMessaging
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.Profile
 
 @Configuration
+@Profile("!test")
 class FcmConfig(
     private val fcmProperty: FcmProperty
 ) {
