+++ src/main/resources/application-test.yaml
@@ -27,14 +27,14 @@ jwt:
     refresh_token_expiration_times: 1209600000 # 2weeks = 1000(=1s) * 60 * 60 * 24 * 14
 
 fcm:
-    type: service_account
-    project_id: yappu-world
-    private_key_id: private-key-id
-    private_key: private-key
-    client_email: client-email
-    client_id: client-id
-    auth_uri: https://accounts.google.com/o/oauth2/auth
-    token_uri: https://oauth2.googleapis.com/token
-    auth_provider_x509_cert_url: https://www.googleapis.com/oauth2/v1/certs
-    client_x509_cert_url: fcm-x509-cert-url
-    universe_domain: googleapis.com
+    type: test
+    project_id: test
+    private_key_id: test
+    private_key: test
+    client_email: test
+    client_id: test
+    auth_uri: test
+    token_uri: test
+    auth_provider_x509_cert_url: test
+    client_x509_cert_url: test
+    universe_domain: test
