+++ src/main/resources/application.yaml
@@ -5,5 +5,6 @@ spring:
         active: local
         group:
             local: console-logging, fcm-local
+            test: console-logging
             dev: console-logging, file-logging, fcm
             prod: file-logging, fcm
